+++ tooling/cli.rs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## \[1.0.0-beta-rc.2]
+
+- Add missing camelcase rename for config
+  - [bdf7072](https://www.github.com/tauri-apps/tauri/commit/bdf707285e3d307ab083009c274ccb56d5053ff2) fix(cli.rs/info): add missing camelCase rename ([#1505](https://www.github.com/tauri-apps/tauri/pull/1505)) on 2021-04-14
+- Fix `tauri info`
+- Properly detect `yarn` and `npm` versions on windows.
+- Fix a panic caused by a wrong field name in `metadata.json`
+- [71666e9](https://www.github.com/tauri-apps/tauri/commit/71666e9f9cfb5499a727b3f95182e89073f67d7b) fix(cli.rs): fix panic & use `cmd` to run `yarn`&`npm` on windows ([#1511](https://www.github.com/tauri-apps/tauri/pull/1511)) on 2021-04-17
+- Sync `metadata.json` via script to update version reference to cli.js, tauri (core) and tauri-build.
+  - [1f64927](https://www.github.com/tauri-apps/tauri/commit/1f64927362ef20761d7cd3591281519eb292aa33) chore: sync cli.rs metadata.json file versions ([#1534](https://www.github.com/tauri-apps/tauri/pull/1534)) on 2021-04-19
+
 ## \[1.0.0-beta-rc.1]
 
 - Missing the `files` property in the package.json which mean that the `dist` directory was not published and used.